// apps/inventory/src/app/remote-entry/model/purchase.interface.ts
  export interface Reference {
    sel: boolean;
    addItem: boolean;
    voucherID: number;
    vNo: string;
    vDate: string; // You can use Date if you convert the string to a Date object
    referenceNo: string | null;
    accountID: number;
    accountName: string;
    amount: number;
    partyInvNo: string | null;
    partyInvDate: string | null; // You can use Date if you convert the string to a Date object
    id: number;
    voucherType: string;
    mobileNo: string | null;
  }